feat(routing): #112(1/4) tripId-scoped 라우트 일반화
`/trip/[tripId]/{map,list,schedule,gmaps-import,items}` 도입.
기존 `/map`, `/list`, `/schedule`, `/gmaps-import`, `/items/*` 는 활성 trip 으로 리디렉트.
TripIdProvider 가 layout 에서 멤버십을 검증하고 클라이언트에 tripId 를 주입한다.

- lib/data.ts / API: tripId 파라미터 수용 (쿼리: ?tripId=)
- useItems: 컨텍스트 tripId 를 API 호출에 전달
- Navigation/FAB/ItemList/ItemForm: useTripPath 로 URL 빌드
- 접근 권한 없는 trip 진입 시 TripAccessDenied 친절 페이지
