+++ app/api/items/[id]/route.ts
@@ -86,9 +86,15 @@ function validatePartial(body: Record<string, unknown>): string | null {
 
 type RouteContext = { params: { id: string } }
 
-export async function GET(_req: NextRequest, { params }: RouteContext) {
+function getTripIdFromRequest(request: NextRequest): string | null {
+  const v = request.nextUrl.searchParams.get('tripId')
+  return v && v.trim() ? v : null
+}
+
+export async function GET(request: NextRequest, { params }: RouteContext) {
   const client = createRouteHandlerSupabase()
-  const items = await readItems(client)
+  const tripId = getTripIdFromRequest(request)
+  const items = await readItems(client, tripId)
   const item = items.find(i => i.id === params.id)
   if (!item) {
     return NextResponse.json({ error: '항목을 찾을 수 없습니다.' }, { status: 404 })
@@ -98,7 +104,8 @@ export async function GET(_req: NextRequest, { params }: RouteContext) {
 
 export async function PUT(request: NextRequest, { params }: RouteContext) {
   const client = createRouteHandlerSupabase()
-  const items = await readItems(client)
+  const tripId = getTripIdFromRequest(request)
+  const items = await readItems(client, tripId)
   const idx = items.findIndex(i => i.id === params.id)
   if (idx === -1) {
     return NextResponse.json({ error: '항목을 찾을 수 없습니다.' }, { status: 404 })
@@ -124,21 +131,22 @@ export async function PUT(request: NextRequest, { params }: RouteContext) {
   }
 
   items[idx] = updated
-  await writeItems(client, items)
+  await writeItems(client, items, tripId)
 
   return NextResponse.json({ item: updated })
 }
 
-export async function DELETE(_req: NextRequest, { params }: RouteContext) {
+export async function DELETE(request: NextRequest, { params }: RouteContext) {
   const client = createRouteHandlerSupabase()
-  const items = await readItems(client)
+  const tripId = getTripIdFromRequest(request)
+  const items = await readItems(client, tripId)
   const idx = items.findIndex(i => i.id === params.id)
   if (idx === -1) {
     return NextResponse.json({ error: '항목을 찾을 수 없습니다.' }, { status: 404 })
   }
 
   items.splice(idx, 1)
-  await writeItems(client, items)
+  await writeItems(client, items, tripId)
 
   return NextResponse.json({ ok: true })
 }
