+++ app/map/page.tsx
@@ -1,12 +1,18 @@
-'use client'
+import { redirectToActiveTrip } from '@/lib/activeTripRedirect'
 
-import { Suspense } from 'react'
-import PlanScreen from '@/components/Plan/PlanScreen'
+export default async function LegacyMapRedirect({
+  searchParams,
+}: {
+  searchParams: Record<string, string | string[] | undefined>
+}) {
+  await redirectToActiveTrip('map', buildQuery(searchParams))
+}
 
-export default function MapPage() {
-  return (
-    <Suspense fallback={null}>
-      <PlanScreen basePath="/map" />
-    </Suspense>
-  )
+function buildQuery(searchParams: Record<string, string | string[] | undefined>): string {
+  const sp = new URLSearchParams()
+  for (const [k, v] of Object.entries(searchParams)) {
+    if (Array.isArray(v)) v.forEach(x => sp.append(k, x))
+    else if (v !== undefined) sp.set(k, v)
+  }
+  return sp.toString()
 }
